requirements.txt
accelerate==0.24.1
annotated-types==0.6.0
anyio==3.7.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
diffusers==0.21.4
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.9.0
fsspec==2023.4.0
ftfy==6.1.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.3.0
psutil==5.9.6
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
safetensors==0.4.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.23.2
watchfiles==0.21.0
wcwidth==0.2.9
webcolors==1.3
websockets==12.0
zipp==3.17.0
